When possible, switch to IN or NOT IN rather than multiple OR'd clauses
DBIx::SearchBuilder 1.63 began allowing the IN and NOT IN operators; RT
4.0 could not immediately make use of them because it did depend on a
high enough version of DBIx::SearchBuilder.  RT 4.2, however, depends on
DBIx::SearchBuilder 1.65, and thus can use IN freely.  Do so, as query
optimizers are likely to be able to make better plans based on IN
queries, rather than multiple OR's.

This also rewrites two queries in GetForwardAttachments into one, using
a simple JOIN.