Store the SQLite database under the temporary directory
This must be done _after_ the initial RT_SiteConfig write, because the
DatabaseType is in RT_Config.pm, and thus only known after both
RT_Siteconfig and RT_Config have been loaded.  This necessitates moving
set_config_wrapper to before the call to bootstrap_db, and valling
WriteSet explicitly, as the override used in Set only catches calls to
Set() from .t files, not from RT::Test -- and the change must be written
out, or rt-validator will not find the database in the correct location.